Updating to Asp.net core 3.0
There are somethings with the Okta Integration, they are not on  asp.net core 3.0.  So we manage some fanagling to get this to work for now.  Once the lib is updated the package will be updated to reflect that.  (Or i figure out using ODIC with okta, but so far that has been a fruitless endevor.)

Also appropriately using MVC rather than just pages.
Search by type is removed for now.  need to figure out getting that into the layout.

-Note for future Chad:
Check to ensure you dont need asp.net core 2.0 installed on machines running this.